[GreenPad-vc4] compromise with both old and new VC for `Min()` usage here

+++ GreenPad-vc4/editwing/ip_cursor.cpp
@@ -157,11 +157,11 @@ void Cursor::Redraw( const VPos& s, const VPos& e )
 	}
 	else
 	{
-		RECT rc = { Max(LFT,sp.x), Max(TOP,sp.y), RHT, Min(BTM,sp.y+view_.fnt().H()) };
+		RECT rc = { Max(LFT,sp.x), Max(TOP,sp.y), RHT, Min(BTM,(int)(sp.y+view_.fnt().H())) };
 		::InvalidateRect( caret_->hwnd(), &rc, FALSE );
-		RECT re = { LFT, Max(TOP,ep.y), Min(RHT,ep.x), Min(BTM,ep.y+view_.fnt().H()) };
+		RECT re = { LFT, Max(TOP,ep.y), Min(RHT,ep.x), Min(BTM,(int)(ep.y+view_.fnt().H())) };
 		::InvalidateRect( caret_->hwnd(), &re, FALSE );
-		RECT rd = { LFT, Max(TOP,rc.bottom), RHT, Min((long)BTM,re.top) };
+		RECT rd = { LFT, Max(TOP,rc.bottom), RHT, Min(BTM,(int)re.top) };
 		::InvalidateRect( caret_->hwnd(), &rd, FALSE );
 	}
 }
